+++ src/themes/teams/utils/index.ts
@@ -1,7 +1,0 @@
-import { ThemeInput } from '../../types'
-
-export const getSideArrow = (theme: ThemeInput) => {
-  const { rtl, siteVariables } = theme
-  const { arrowLeft, arrowRight } = siteVariables
-  return rtl ? arrowLeft : arrowRight
-}